requirements.txt
streamlit 
google-cloud-speech 
google-auth 
plotly 
pandas
pydub